Non-vanilla history entry style "fix" for NROC.txt
This change is solely one of syntax style.  This short non-vanilla
file is the only place in the codebase that uses the prior single-
line history entry syntax style, which apparently/presumably is still
valid for the engine.

However, it's not for my SWMH character history parser, and though I
could change the code to deal with arbitrary permutations of this style
type with some refactoring work (that I'd rather not do), it was so
rare that it made more sense to just "fix" the offending history entries
in this file.  The "culture melter" tool now clears the entire SWMH
character history database with this applied.