Scripted changes to add missing history files
So, SWMH imports a lot of vanilla character history files. This is
normally fine, but recently we're running into issues where SWMH and
vanilla have char id conflicts, because a file SWMH doesn't override
shares an id with an SWMH file. This change generates every file that
needs to be overridden, and adds the vanilla characters that aren't
dupes to the overwriting files. This should ONLY fix bugs and not add
any new characters whatsoever.